Refactor (#27)
* Refactor...
Moved XDSUser model to users app
Moved permission checker class to users app
Moved password validators to users app
Moved XDSConfiguration to configurations app
Moved XDSUIConfiguration to configurations app
Various DB migrations
Moved tests to relevant apps
Updated XDS-API views to be class based

* Added auth to es-api...
Switch es-api to class based views from function based
Update OPEN_ENDPOINTS and PermissionsChecker to allow RegEx matching
Fix to how PermissionsChecker determines module name
Fix for permissions generation
Update tests to add correct permission
Updates to formatting to meet flake8 standard
Update to GetSpotlightCourses now returns nothing if none active
Move CourseInformationMapping to configurations

* Updated imports to match isort standard
