+++ app/core/tests/test_models_unit.py
@@ -1,30 +1,15 @@
+from configurations.models import XDSConfiguration
+from core.models import (CourseDetailHighlight, CourseSpotlight, Experience,
+                         InterestList, SearchFilter, SearchSortOption,
+                         XDSUIConfiguration)
 from django.test import tag
-
-from core.models import (CourseDetailHighlight, CourseInformationMapping,
-                         CourseSpotlight, Experience, InterestList,
-                         SearchFilter, SearchSortOption, XDSConfiguration,
-                         XDSUIConfiguration, XDSUser)
+from users.models import XDSUser
 
 from .test_setup import TestSetUp
 
 
 @tag('unit')
 class ModelTests(TestSetUp):
-
-    def test_create_xds_configuration(self):
-        """Test that creating a new XDS Configuration entry is successful\
-        with defaults """
-        xdsConfig = XDSConfiguration(target_xis_metadata_api="test")
-
-        self.assertEqual(xdsConfig.target_xis_metadata_api, "test")
-
-    def test_create_xds_ui_configuration(self):
-        """Test that creating a new XDSUI Configuration is successful with \
-            defaults"""
-        config = XDSConfiguration(target_xis_metadata_api="test")
-        uiConfig = XDSUIConfiguration(xds_configuration=config)
-
-        self.assertEqual(uiConfig.search_results_per_page, 10)
 
     def test_create_search_filter(self):
         """Test that creating a search filter object works correctly"""
@@ -72,28 +57,6 @@ class ModelTests(TestSetUp):
 
         self.assertEqual(c_id, spotlight.course_id)
         self.assertTrue(spotlight.active)
-
-    def test_create_courseInformationMapping(self):
-        """Tests the creation of a course information object"""
-
-        config = XDSConfiguration(target_xis_metadata_api="test")
-        uiConfig = XDSUIConfiguration(xds_configuration=config)
-
-        # course mappings
-        course_title = 'Course.TestTitle'
-        course_description = 'Course.TestDescription'
-        course_url = 'Course.TestUrl'
-
-        courseInformation = CourseInformationMapping(
-            xds_ui_configuration=uiConfig,
-            course_title=course_title,
-            course_description=course_description,
-            course_url=course_url)
-
-        self.assertEqual(courseInformation.course_title, course_title)
-        self.assertEqual(courseInformation.course_description,
-                         course_description)
-        self.assertEqual(courseInformation.course_url, course_url)
 
     def test_create_experience(self):
         """Tests that creating a course is successful"""
