+++ app/es_api/utils/queries.py
@@ -2,11 +2,10 @@ import json
 import logging
 import os
 
+from configurations.models import XDSConfiguration
+from core.models import CourseSpotlight, SearchFilter, SearchSortOption
 from elasticsearch_dsl import A, Document, Q, Search, connections
 from elasticsearch_dsl.query import MoreLikeThis
-
-from core.models import (CourseSpotlight, SearchFilter, SearchSortOption,
-                         XDSConfiguration)
 
 connections.create_connection(alias='default',
                               hosts=[os.environ.get('ES_HOST'), ], timeout=60)
